+++ app/src/main/java/woowacourse/shopping/data/repository/AuthRepository.kt
@@ -1,23 +1,11 @@
 package woowacourse.shopping.data.repository
 
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.first
-import woowacourse.shopping.data.localdata.UserDataStore
+import kotlinx.coroutines.flow.StateFlow
 
-class AuthRepository(
-    private val userDataStore: UserDataStore,
-) {
-    private val _userName = MutableStateFlow<String?>(null)
-    private val _password = MutableStateFlow<String?>(null)
+interface AuthRepository {
+    val userName: StateFlow<String?>
 
-    val userName = _userName.asStateFlow()
-    val password = _password.asStateFlow()
+    val password: StateFlow<String?>
 
-    suspend fun loadCredentialsToMemory() {
-        val credentials = userDataStore.userCredentialsFlow.first()
-
-        _userName.value = credentials.username
-        _password.value = credentials.password
-    }
+    suspend fun loadCredentialsToMemory()
 }
